[node2nix] nix-shell PATH
It seems the node modules are not in PATH for the nix-shell environment.
I guess this makes sense for libraries that are meant to be used by node
(ensuring the javascript runtime can locate dependencies is not the same
as ensuring the interactive shell (e.g. Bash) can locate them).

I suspect the right way to fix this is to add any such node packages
(those which should expose executables to the interactive shell) as
buildInputs (or propogatedBuildInputs?) in the nix derivation that
nix-shell uses.

As a hacky workaround/proof-of-concept, I tried manipulating PATH while
inside the nix-shell:

  PATH=/nix/store/xxxxxx-node-dependencies-hello-world-1.0.0/lib/node_modules/.bin:$PATH

This allowed me to get to the next error:

  This dependency was not found:

  * webpack/hot/dev-server in multi (webpack)-dev-server/client?http://localhost:8080 webpack/hot/dev-server ./src/main.js

Full output below:

  [nix-shell:~/Development/code/vue/hello-world]$ npm run dev

  > hello-world@1.0.0 dev /home/ivan/Development/code/vue/hello-world
  > webpack-dev-server --inline --progress --config build/webpack.dev.conf.js

   94% asset optimization

   ERROR  Failed to compile with 2 errors                                                                                                                               08:39:51

  This dependency was not found:

  * webpack/hot/dev-server in multi (webpack)-dev-server/client?http://localhost:8080 webpack/hot/dev-server ./src/main.js

  To install it, you can run: npm install --save webpack/hot/dev-server

  This relative module was not found:

  * ./src/main.js in multi (webpack)-dev-server/client?http://localhost:8080 webpack/hot/dev-server ./src/main.js