ln -s $NODE_PATH node_modules
As a proof of concept, I tried the following:

  git rm supplement.*
  nix-shell \
      --packages nodePackages.node2nix \
      --run 'node2nix --development -8 -l package-lock.json'
  nix-shell -A shell

Then, in the nix-shell, I ran:

  ln -s $NODE_PATH node_modules
  npm run dev

This succeeded, and I was able to visit localhost:8080 to verify the
server was running.

This is not a robust solution, but at least clarifies the issues we're
up against. Relevant discussions mentioning this workaround:
  https://github.com/svanderburg/node2nix/issues/63
  https://github.com/svanderburg/node2nix/issues/29

Discussion of a different solution I want to try:
  https://github.com/svanderburg/node2nix/issues/51